Import known civs and war statuses from SAVs, and add a diplomacy popup
We now have the lower right hand "D" button for opening the diplomacy popup, and the popup lists the civs we know and whether we're at war with them.

In the LEAD struct I changed some byte arrays to longs/ints to make it easier to dump them to the console when debugging.

In order for the "D" button to work I had to shuffle the "GameStatus" node in front of the unit buttons node, because the invisible HBox for the unit buttons was intercepting all the mouse events.

#571